Adjustor cmpt moved to its own folder
In order to clean up and be consistent with recommandations
