+++ src/app/adjustor/adjustor-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { AdjustorComponent } from '../adjustor/adjustor.component';
+import { AdjustorComponent } from './adjustor/adjustor.component';
 
 const routes: Routes = [
 {
