+++ src/app/adjustor/adjustor/adjustor.component.css
