+++ src/app/adjustor/adjustor/adjustor.component.scss
